+++ how-to/workspace-platform-starter/client/types/module/shapes/lifecycle-shapes.d.ts
@@ -16,7 +16,8 @@ export type LifecycleEvents =
 	| "workspace-changed"
 	| "page-changed"
 	| "apps-changed"
-	| "favorite-changed";
+	| "favorite-changed"
+	| "condition-changed";
 /**
  * The type for a lifecycle event handler.
  */
@@ -110,4 +111,13 @@ export interface ThemeChangedLifecyclePayload {
 	 * The payload.
 	 */
 	payload?: CustomPaletteSet;
+}
+/**
+ * Condition changed event payload.
+ */
+export interface ConditionChangedLifecyclePayload {
+	/**
+	 * The condition that changed, or empty to determine many might have changed.
+	 */
+	conditionId?: string;
 }